fix xgboost test flake (#70)
Currently, xgboost test will flake because we infer the number of
classes from training data. Since the number of rows is small, it's
possible that some class does not appear in the targets in training data
but appear in the targets in validation data, which will cause a test
flake.