Fix `ResNet`'s `in_channels` calculations (#220)
Current test passed because we always have target column and
`len(col_stats) - 1` happens to be the same as the target column. This
no longer holds when we need to handle a table with no target column.
This PR fixes the issue by computing the number of columns based on
`col_names_dict`.