Don't put assertions on N/A imputation correctness (#433)
I understand the intention of the assertion introduced in #307, however,
we should not check it at runtime as it causes a D2H sync. It should
rather be covered in tests, and I believe there's no need to add new
tests as existing test cases already cover the correctness of the N/A
imputation.